+++ GPU/GLES/FramebufferManagerGLES.h
@@ -64,8 +64,6 @@ public:
 	void DeviceRestore(Draw::DrawContext *draw);
 
 protected:
-	void SetViewport2D(int x, int y, int w, int h) override;
-
 	// Used by ReadFramebufferToMemory and later framebuffer block copies
 	void BlitFramebuffer(VirtualFramebuffer *dst, int dstX, int dstY, VirtualFramebuffer *src, int srcX, int srcY, int w, int h, int bpp) override;
