+++ GPU/GPU.vcxproj.filters
@@ -285,6 +285,9 @@
     <ClInclude Include="Software\RasterizerRectangle.h">
       <Filter>Software</Filter>
     </ClInclude>
+    <ClInclude Include="Common\PresentationCommon.h">
+      <Filter>Common</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
@@ -565,6 +568,9 @@
     </ClCompile>
     <ClCompile Include="Software\RasterizerRectangle.cpp">
       <Filter>Software</Filter>
+    </ClCompile>
+    <ClCompile Include="Common\PresentationCommon.cpp">
+      <Filter>Common</Filter>
     </ClCompile>
   </ItemGroup>
 </Project>
